integration: extract test_prepared_partitioner()
The test is extracted into `statements/prepared.rs`.
