+++ scylla/tests/integration/session/session.rs
@@ -1,13 +1,11 @@
 use crate::utils::{
-    create_new_session_builder, scylla_supports_tablets, setup_tracing, supports_feature,
-    unique_keyspace_name, PerformDDL,
+    create_new_session_builder, setup_tracing, supports_feature, unique_keyspace_name, PerformDDL,
 };
 
 use scylla::client::execution_profile::ExecutionProfile;
 use scylla::client::session::Session;
 use scylla::errors::OperationType;
 use scylla::errors::{DbError, ExecutionError, RequestAttemptError};
-use scylla::routing::partitioner::PartitionerName;
 use scylla::routing::Token;
 use scylla::serialize::row::SerializeRow;
 use scylla::statement::prepared::PreparedStatement;
@@ -362,62 +360,6 @@ async fn test_named_bind_markers() {
     for wrongmap in wrongmaps {
         assert!(session.execute_unpaged(&prepared, &wrongmap).await.is_err());
     }
-}
-
-#[tokio::test]
-async fn test_prepared_partitioner() {
-    let session = create_new_session_builder().build().await.unwrap();
-    let ks = unique_keyspace_name();
-
-    // This test uses CDC which is not yet compatible with Scylla's tablets.
-    let mut create_ks = format!(
-        "CREATE KEYSPACE IF NOT EXISTS {ks} WITH REPLICATION = {{'class' : 'NetworkTopologyStrategy', 'replication_factor' : 1}}");
-    if scylla_supports_tablets(&session).await {
-        create_ks += " AND TABLETS = {'enabled': false}"
-    }
-
-    session.ddl(create_ks).await.unwrap();
-    session.use_keyspace(ks, false).await.unwrap();
-
-    session
-        .ddl("CREATE TABLE IF NOT EXISTS t1 (a int primary key)")
-        .await
-        .unwrap();
-
-    session.await_schema_agreement().await.unwrap();
-    session.refresh_metadata().await.unwrap();
-
-    let prepared_statement_for_main_table = session
-        .prepare("INSERT INTO t1 (a) VALUES (?)")
-        .await
-        .unwrap();
-
-    assert_eq!(
-        prepared_statement_for_main_table.get_partitioner_name(),
-        &PartitionerName::Murmur3
-    );
-
-    if option_env!("CDC") == Some("disabled") {
-        return;
-    }
-
-    session
-        .ddl("CREATE TABLE IF NOT EXISTS t2 (a int primary key) WITH cdc = {'enabled':true}")
-        .await
-        .unwrap();
-
-    session.await_schema_agreement().await.unwrap();
-    session.refresh_metadata().await.unwrap();
-
-    let prepared_statement_for_cdc_log = session
-        .prepare("SELECT a FROM t2_scylla_cdc_log WHERE \"cdc$stream_id\" = ?")
-        .await
-        .unwrap();
-
-    assert_eq!(
-        prepared_statement_for_cdc_log.get_partitioner_name(),
-        &PartitionerName::CDC
-    );
 }
 
 #[tokio::test]
