+++ scylla/tests/integration/statements/prepared.rs
@@ -1,11 +1,12 @@
 use scylla::{
     response::{PagingState, PagingStateResponse},
-    routing::Token,
+    routing::{partitioner::PartitionerName, Token},
     statement::Statement,
 };
 
 use crate::utils::{
-    create_new_session_builder, setup_tracing, unique_keyspace_name, PerformDDL as _,
+    create_new_session_builder, scylla_supports_tablets, setup_tracing, unique_keyspace_name,
+    PerformDDL as _,
 };
 
 #[tokio::test]
@@ -194,4 +195,62 @@ async fn test_prepared_config() {
 
     assert!(prepared_statement.get_is_idempotent());
     assert_eq!(prepared_statement.get_page_size(), 42);
+}
+
+#[tokio::test]
+async fn test_prepared_partitioner() {
+    setup_tracing();
+
+    let session = create_new_session_builder().build().await.unwrap();
+    let ks = unique_keyspace_name();
+
+    // This test uses CDC which is not yet compatible with Scylla's tablets.
+    let mut create_ks = format!(
+        "CREATE KEYSPACE IF NOT EXISTS {ks} WITH REPLICATION = {{'class' : 'NetworkTopologyStrategy', 'replication_factor' : 1}}");
+    if scylla_supports_tablets(&session).await {
+        create_ks += " AND TABLETS = {'enabled': false}"
+    }
+
+    session.ddl(create_ks).await.unwrap();
+    session.use_keyspace(ks, false).await.unwrap();
+
+    session
+        .ddl("CREATE TABLE IF NOT EXISTS t1 (a int primary key)")
+        .await
+        .unwrap();
+
+    session.await_schema_agreement().await.unwrap();
+    session.refresh_metadata().await.unwrap();
+
+    let prepared_statement_for_main_table = session
+        .prepare("INSERT INTO t1 (a) VALUES (?)")
+        .await
+        .unwrap();
+
+    assert_eq!(
+        prepared_statement_for_main_table.get_partitioner_name(),
+        &PartitionerName::Murmur3
+    );
+
+    if option_env!("CDC") == Some("disabled") {
+        return;
+    }
+
+    session
+        .ddl("CREATE TABLE IF NOT EXISTS t2 (a int primary key) WITH cdc = {'enabled':true}")
+        .await
+        .unwrap();
+
+    session.await_schema_agreement().await.unwrap();
+    session.refresh_metadata().await.unwrap();
+
+    let prepared_statement_for_cdc_log = session
+        .prepare("SELECT a FROM t2_scylla_cdc_log WHERE \"cdc$stream_id\" = ?")
+        .await
+        .unwrap();
+
+    assert_eq!(
+        prepared_statement_for_cdc_log.get_partitioner_name(),
+        &PartitionerName::CDC
+    );
 }
